+++ contrib/handlers/rakuten_card.go
@@ -13,7 +13,6 @@ import (
 
 // RakutenCardStatement build a handler for statements of Rakuten Card (楽天カード 明細).
 func RakutenCardStatement(name, pattern string, table Table, notifier bqloader.Notifier) *bqloader.Handler {
-	type contextKey string
 	var monthKey contextKey = "month"
 
 	re := regexp.MustCompile(`enavi(\d+)\(\d+\)`)
