+++ contrib/handlers/util.go
@@ -1,8 +1,1 @@
 package handlers
-
-// Table identifies BigQuery table.
-type Table struct {
-	Project string
-	Dataset string
-	Table   string
-}
